+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [UNRELEASED]
 ### Added
 - test coverage of release integration
+- test coverage of static category configuration
+- test coverage of static category configuration override
+- test coverage of dynamic category configuration
 ### Changed
 - format of static category configuration (no API breakage as the old format was incompatible with releases in any case)
 ### Fixed